+++ gouge/colourcli.py
@@ -1,6 +1,8 @@
 import logging
+from logging import Handler, LogRecord
+from typing import Any, List, Optional
 
-from blessings import Terminal
+import colorama as clr
 
 
 class Simple(logging.Formatter):
@@ -13,16 +15,18 @@ class Simple(logging.Formatter):
     .. note:: This formatter *suppresses* tracebacks by default! Remember that
         is is meant to give a concise, readable output. If you need to see
         tracebacks on the console, you can override this setting useing
-        \*show_exc\*.
+        \\*show_exc\\*.
     """
 
     @staticmethod
-    def basicConfig(show_exc=True, show_threads=False, *args, **kwargs):
+    def basicConfig(show_exc=True, show_threads=False, force_styling=False,
+                    **kwargs):
+        # type: (bool, bool, bool, Any) -> List[Handler]
         '''
         Convenience method to have a one-liner set-up.
 
-        Both *show_exc* and *show_threads* are directly passed to
-        :py:class:`~.Simple`. The remaining *args* and *kwargs* are passed on to
+        *show_exc*, *show_threads* and *force_styling* are directly passed to
+        :py:class:`~.Simple`. The remaining *kwargs* are passed on to
         :py:func:`logging.basicConfig`.
 
         After returning from :py:func:`logging.basicConfig`, it will fetch the
@@ -32,42 +36,51 @@ class Simple(logging.Formatter):
         modified. This is useful if you want to modify the handlers any further
         (for example using :py:class:`~gouge.filters.ShiftingFilter`).
         '''
-        logging.basicConfig(*args, **kwargs)
+        clr.init(strip=(False if force_styling else None))
+        logging.basicConfig(**kwargs)
         root = logging.getLogger()
         output = []
         for handler in root.handlers:
-            if (hasattr(handler, 'stream') and
-                    hasattr(handler.stream, 'name') and
-                    handler.stream.name in ('<stderr>', '<stdout>')):
-                handler.setFormatter(Simple(show_exc, show_threads))
-                output.append(handler)
+            stream = getattr(handler, 'stream', None)
+            if not stream:
+                continue
+
+            stream_name = getattr(stream, 'name', None)
+            if stream_name not in ('<stderr>', '<stdout>'):
+                continue
+
+            handler.setFormatter(Simple(show_exc, show_threads))
+            output.append(handler)
         return output
 
-    def __init__(self, show_exc=False, show_threads=False, *args, **kwargs):
-        logging.Formatter.__init__(self, *args, **kwargs)
+    def __init__(self, show_exc=False, show_threads=False, fmt=None,
+                 datefmt=None, force_styling=False):
+        # type: (bool, bool, Optional[str], Optional[str], Optional[bool]) -> None
+        logging.Formatter.__init__(self, fmt, datefmt)
         self.show_threads = show_threads
         self.show_exc = show_exc
-        self.term = Terminal()
+        self.force_styling = force_styling
 
     def format(self, record):
+        # type: (LogRecord) -> str
         if record.levelno <= logging.DEBUG:
-            colorize = self.term.bold_black
+            colorize = clr.Style.BRIGHT + clr.Fore.BLACK
         elif record.levelno <= logging.INFO:
-            colorize = self.term.cyan
+            colorize = clr.Fore.CYAN
         elif record.levelno <= logging.WARNING:
-            colorize = self.term.yellow
+            colorize = clr.Fore.YELLOW
         elif record.levelno <= logging.ERROR:
-            colorize = self.term.bold_red
+            colorize = clr.Style.BRIGHT + clr.Fore.RED
         else:
-            colorize = self.term.bold_yellow_on_red
+            colorize = clr.Style.BRIGHT + clr.Fore.YELLOW + clr.Back.RED
 
         record.message = record.getMessage()
-        record.asctime = self.formatTime(record, self.datefmt)
+        record.asctime = self.formatTime(record, self.datefmt or '')
 
         message_items = [
-            '{t.green}{asctime}{t.normal}',
-            '{levelcolor}{levelname:<10}{t.normal}',
-            '{t.bold}[{name}]{t.normal}',
+            '{f.GREEN}{asctime}{s.RESET_ALL}',
+            '{levelcolor}{levelname:<10}{s.RESET_ALL}',
+            '{s.BRIGHT}[{name}]{s.RESET_ALL}',
             '{message}',
         ]
         if self.show_threads:
@@ -79,13 +92,17 @@ class Simple(logging.Formatter):
             if record.exc_info:
                 # Cache the traceback text to avoid converting it multiple times
                 # (it's constant anyway)
-                if not record.exc_text:
-                    record.exc_text = self.formatException(record.exc_info)
+                exc_text = getattr(record, 'exc_text', '')
+                if not exc_text:
+                    record.exc_text = self.formatException(  # type: ignore
+                        record.exc_info)
 
-            if record.exc_text:
-                message_template += '\n{t.red}{exc_text}{t.normal}'
+            exc_text = getattr(record, 'exc_text', '')
+            if exc_text:
+                message_template += '\n{f.RED}{exc_text}{s.RESET_ALL}'
 
         return message_template.format(
-            t=self.term,
             levelcolor=colorize,
+            f=clr.Fore,
+            s=clr.Style,
             **vars(record))