+++ docs/changelog.rst
@@ -1,6 +1,19 @@
 Changelog
 =========
 
+Version 1.3.0
+-------------
+
+* Replace ``blessings`` with ``colorama``. This will make colors work on
+  Windows as well.
+
+
+Version 1.2.1
+-------------
+
+* Add type-hints & fix minor typing bugs
+* Add ``gouge.__version__`` attribute
+
 Version 1.2.0
 -------------